tmbu/requirements.txt
Django==5.0.1
psycopg[binary]==3.2.3
nltk==3.9.1
openai==1.59.6
pymilvus==2.5.10
python-dotenv==1.0.1
sentence_transformers==3.3.0
stripe==11.2.0
transformers==4.46.2
safetensors==0.4.5
gunicorn==23.0.0
whitenoise==6.8.2
# django-apscheduler==0.7.0
django-recaptcha==4.1.0
requests==2.32.3
celery==5.5.3
django-celery-results==2.5.1
django-allauth==65.11.0
PyJWT==2.10.1
cryptography==46.0.1




